backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]>=0.30.0
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.9.0
sse-starlette==2.1.0
numpy>=2.1.0
scikit-learn>=1.5.2
tenacity==8.5.0
websockets>=13.0.0
